requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
aiosqlite==0.20.0

# Testing
pytest==8.0.0
pytest-asyncio==0.21.1
